Use consistent file names for DTOs, refactor
