+++ lib/src/main/java/com/thenewboston/data/dto/bankapi/blockdto/Blocks.kt
@@ -1,4 +1,4 @@
-package com.thenewboston.data.dto.bankapi.banktransactiondto
+package com.thenewboston.data.dto.bankapi.blockdto
 
 import com.thenewboston.utils.DateSerializer
 import kotlinx.datetime.LocalDateTime
@@ -7,26 +7,40 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class BlockList(
+
+    @SerialName("count")
     val count: Long,
+
+    @SerialName("next")
     val next: String?,
+
+    @SerialName("previous")
     val previous: String?,
-    val results: List<Block>
+
+    @SerialName("results")
+    val blocks: List<Block>
 )
 
 @Serializable
 data class Block(
+
     @SerialName("id")
     val id: String,
+
     @Serializable(with = DateSerializer::class)
     @SerialName("created_date")
     val createdDate: LocalDateTime,
+
     @Serializable(with = DateSerializer::class)
     @SerialName("modified_date")
     val modifiedDate: LocalDateTime?,
+
     @SerialName("balance_key")
     val balanceKey: String,
+
     @SerialName("sender")
     val sender: String,
+
     @SerialName("signature")
     val signature: String
 )
