+++ lib/src/main/java/com/thenewboston/data/dto/bankapi/invalidblockdto/request/PostInvalidBlockRequest.kt
@@ -1,0 +1,4 @@
+package com.thenewboston.data.dto.bankapi.invalidblockdto.request
+
+class PostInvalidBlockRequest {
+}
